requirements.txt
Flask==2.2.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
geoalchemy2==0.12.5
psycopg2-binary==2.9.4
